Fix the Property index issue. Some refactoring.
