+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/datamodel/property/Property.scala
@@ -8,9 +8,8 @@ package edu.illinois.cs.cogcomp.saul.datamodel.property
 
 import java.util
 
-import edu.illinois.cs.cogcomp.lbjava.classify.{Classifier, FeatureVector}
+import edu.illinois.cs.cogcomp.lbjava.classify.{ Classifier, FeatureVector }
 import edu.illinois.cs.cogcomp.saul.datamodel.node.Node
-import edu.illinois.cs.cogcomp.saul.datamodel.property.features.ClassifierContainsInLBP
 
 import scala.reflect.ClassTag
 
@@ -21,7 +20,7 @@ import scala.reflect.ClassTag
   */
 trait Property[T] {
 
-  val containingPackage = "LBP_Package"
+  private[property] val containingPackage = "LBP_Package"
   val name: String
 
   val tag: ClassTag[T]
@@ -35,35 +34,13 @@ trait Property[T] {
 
   def outputType: String = "discrete"
 
+  def allowableValues: Array[String] = Array.empty[String]
+
   def compositeChildren: Option[util.LinkedList[Classifier]] = None
 }
 
 object Property {
 
   /** Transfer a properties to a lbj classifier. */
-  def makeClassifier[T](property: Property[T]): Classifier = {
-    new ClassifierContainsInLBP {
-      name = property.name
-      containingPackage = property.containingPackage
-
-      override def getOutputType: String = property.outputType
-
-      override def classify(o: scala.Any): FeatureVector = property.featureVector(o.asInstanceOf[T])
-
-      override def discreteValue(o: scala.Any): String = property.featureVector(o.asInstanceOf[T]).discreteValueArray().head
-
-      override def realValue(o: scala.Any): Double = property.featureVector(o.asInstanceOf[T]).realValueArray().head
-
-      override def discreteValueArray(o: scala.Any): Array[String] = property.featureVector(o.asInstanceOf[T]).discreteValueArray()
-
-      override def realValueArray(o: scala.Any): Array[Double] = property.featureVector(o.asInstanceOf[T]).realValueArray()
-
-      override def getCompositeChildren: util.LinkedList[_] = property.compositeChildren.orNull
-    }
-  }
-
-  def addToFeatureVector[T](property: Property[T], instance: T, fv: FeatureVector): FeatureVector = {
-    fv.addFeatures(property.featureVector(instance))
-    fv
-  }
+  def convertToClassifier[T](property: Property[T]): Classifier = new LBPClassifier[T](property)
 }
